requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file> -c pytorch
# platform: linux-64
cudatoolkit=11.3.1
python=3.8.13
pytorch=1.12.1
tensorboard=2.8.0
numpy=1.23.1